server/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
asn1crypto==1.5.1
atpublic==4.1.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cryptography==42.0.8
Faker==25.8.0
filelock==3.15.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
platformdirs==4.2.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
snowflake==0.9.0
snowflake-connector-python==3.10.1
snowflake-id==1.0.2
snowflake-snowpark-python==1.18.0
snowflake._legacy==0.9.0
snowflake.core==0.9.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
SQLAlchemy-serializer==1.4.12
tomlkit==0.12.5
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.1
